requirements.txt
alembic==1.15.1
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==3.0.10
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2025.1
requests==2.26.0
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
Werkzeug==3.1.3
psycopg2-binary

# requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.2
Flask==2.1.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
